pyproject.toml
[project]
name = "agentictutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.49.0",
    "python-dotenv>=1.1.0",
]
